requirements.txt
pytest
pytest-html
pytest-sugar
pytest-rerunfailures
pytest-ordering
allure-pytest
uiautomator2
requests
pillow